+++ NEWS.md
@@ -1,4 +1,4 @@
-# crane 0.3.1.9039
+# crane 0.3.1.9040
 
 ## New Functions and Functionality
 
@@ -23,13 +23,15 @@
 
 * `g_lineplot()`, `g_lineplot_table()`, and `preprocess_lineplot_data()` were deprecated. (#184)
 
-*  Added `tbl_coxph()` which computes the pairwise cox proportional hazards test and returns a gtsummary object (#159)
+* Added `tbl_coxph()` to render pairwise Cox proportional hazards results as a gtsummary table. (#159)
 
 * Added `tbl_hierarchical_incidence_rate()` which computes the incidence rate of adverse events (#211)
 
 * Added `adjust_stat_columns_wrap()` which keep the presentation of statistics in one row (#219)
 
-* `get_cox_pairwise_df` now allows for different methods to compute p-value (#191) and to handle ties (#193)
+* `get_cox_pairwise_df()` now allows for different methods to compute p-value (#191) and to handle ties (#193).
+
+* `tbl_coxph()` now accepts a data.frame created by `get_cox_pairwise_df()` as input. (#207)
 
 * Added `add_grade_column()` to inject a grade-label column into `tbl_hierarchical_rate_by_grade()` output. Decoupled from the table builder to prevent Cartesian join explosion in `tbl_merge()`. (#226)